backend/pyproject.toml
[tool.poetry]
name = "my-company-consent-demo-app"
version = "1.0.0"
description = "MyCompany consent demo app backend"
authors = ["Digital Living International Ltd"]
license = "BSD-3-Clause"
packages = [{include = "app", from="."}]

[tool.poetry.dependencies]
python = "^3.9"
fastapi = "^0.68.1"
uvicorn = "^0.15.0"
aiofiles = "^0.4.0"
invoke = "^1.4.1"
httpx = "^0.19.0"
mock = "^4.0.2"
python-multipart = "^0.0.5"
pyjwt = "^2.1.0"
pydantic = {extras = ["dotenv"], version = "^1.8.2"}
cryptography = "^3.1"
firedantic = "^0.2.8"
loguru = "^0.5.3"
stringcase = "^1.2.0"
pyjwt-key-fetcher = "^0.1.0"

[tool.poetry.dev-dependencies]
isort = "^5.5.3"
bandit = "^1.7.0"

[tool.isort]
profile = "black"

[build-system]
requires = ["poetry>=1.0.0"]
build-backend = "poetry.masonry.api"
